rename mysql.ts to properties.ts
