+++ src/components/sections/Modal/MarkdownEditor.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { useState } from 'react'
 import Markdown from 'react-markdown'
 import { Task } from '../../../util/types'
-import { DESCRIPTION_COLUMN } from '../../../util/mysql'
+import { DESCRIPTION_COLUMN } from '../../../util/properties'
 import useWindowDimensions from '../../../util/util'
 import Button from '../../buttons/Button'
 import ToastMessage from '../../util/ToastMessage'
