+++ src/components/sections/Modal/TaskCardModal.tsx
@@ -8,7 +8,7 @@ import MarkdownEditor from './MarkdownEditor'
 import EpicTaskList from './EpicTaskList'
 import EpicGenerateTasksButton from './EpicGenerateTasksButton'
 import { useSearchParams } from 'react-router-dom'
-import { NAME_COLUMN } from '../../../util/mysql'
+import { NAME_COLUMN } from '../../../util/properties'
 import { Task } from '../../../util/types'
 import Button from '../../buttons/Button'
 import { route } from '../../../util/queryRouting'
