+++ src/util/properties.ts
